[chore #159727523] Refine code base structure
