--- app/home/templates/index.html
@@ -1,133 +1,0 @@
-{% extends 'base_frontend.html' %}
-
-{% block body %}
-<div class="container headerbg">
-    <div class="wrappers">
-        <div class="col-5 mt-">
-            <h1 class="title">Questions API endpoints</h1>
-            <!-- link -->
-            <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    [<b class="text-primary">GET</b>]Fetch all questions 
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/questions">/api/v1/questions</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-            <!-- link -->
-            <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    [<b class="text-primary">GET</b>]Fetch a specific question  
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/questions/1">/api/v1/questions/1</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-            <!-- link -->
-            <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    <b class="text-primary">POST</b> /questions   
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/questions">/api/v1/questions</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-            <!-- link -->
-            <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    <b class="text-primary">PUT</b> /questions   
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/questions">/api/v1/questions</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-             <!-- link -->
-             <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    <b class="text-primary">DELETE</b> /questions   
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/questions/1">/api/v1/questions/1</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-            <!-- link -->
-            <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    <b class="text-primary">POST</b> /question answer   
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/questions/1/answer">/api/v1/questions/1/answer</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-        </div>
-        <div class="col-5 pull-right">
-            <h1 class="title">Users API endpoints</h1>
-            <!-- link -->
-            <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    [<b class="text-primary">GET</b>]Fetch all users 
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/users">/api/v1/users</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-            <!-- link -->
-            <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    [<b class="text-primary">GET</b>]Fetch a specific question  
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/users/1">/api/v1/users/1</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-            <!-- link -->
-            <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    <b class="text-primary">POST</b> /users   
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/users">/api/v1/users</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-            <!-- link -->
-            <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    <b class="text-primary">PUT</b> /users   
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/users">/api/v1/users</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-             <!-- link -->
-             <div class="card p-15  mb-0" >
-                <div class="pull-left">
-                    <b class="text-primary">DELETE</b> /users   
-                </div>
-                <div class="pull-right hatch">
-                    <a href="/api/v1/users/1">/api/v1/users/1</a>
-                </div>                
-            </div>
-            <!-- ./link -->
-        </div>
-    </div>
-    <!-- ./wrapper -->
-</div>
-<!-- ./container -->
-<div class="clearfix"></div>
-<div class="full-span bg-primary p-15 text-center">
-    <div class="mega-text stretchLeft">Ask A Question</div>
-    <p class="stretchLeft">
-        You can ask a question / post on our community by providing all the required details.
-    </p>
-</div>
-{% endblock %}
