--- app/questions/views.py
@@ -1,3 +1,0 @@
-#################
-# imports
-#################
